agmi/agmi-video-generator/requirements.txt
google-genai
moviepy
requests
beautifulsoup4

